F_T30603 move DocumentBundle Exceptions to core
- adjust use statements
